Run cypress in container with fixed version of Chrome 79
We are now again seeing builds on GitHub Actions stalling because
Cypress cannot connect to Chrome. This seems to be an issue from
Chrome 80 and on.

The issue might be fixed by upgrading to Cypress 4.

For now we simply run Chrome from a Docker image provided by Cypress 
with a fixed version.

See https://github.com/cypress-io/github-action#docker-image